move `mainPublisher`, `mainSubscriber` to `scripts`
