+++ services/claim-relayer/scripts/mainPublisher.ts
