+++ services/claim-relayer/scripts/mainSubscriber.ts
